+++ src/main/java/com/oghuz/engine/V8Engine.java
@@ -1,7 +1,7 @@
 package com.oghuz.engine;
 
 public class V8Engine extends Engine {
-    public V8Engine(engineType type, Short fuelCapacity, Short cylinderCount) {
+    public V8Engine(engineType type, Double fuelCapacity, Short cylinderCount) {
         super(type, fuelCapacity, cylinderCount);
         type = engineType.vType;
         fuelCapacity = 6.2;
